+++ src/main/java/com/talentwunder/financetracker/model/response/TransactionSummaryResponse.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -12,8 +13,8 @@ import java.util.List;
 @NoArgsConstructor
 @Builder
 public class TransactionSummaryResponse {
-    private Double balance;
-    private Double totalIncome;
-    private Double totalExpense;
+    private BigDecimal balance;
+    private BigDecimal totalIncome;
+    private BigDecimal totalExpense;
     private List<TransactionResponse> transactions;
 }
